Revert "Oursource search to component"
This reverts commit 5c9f709b7d92bc882e53973d5fcd6d7aae98eca9.
